Submit message only show for essay-containing quiz, rearranging source code
